requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.2
    # via langchain-community
aiosignal==1.4.0
    # via aiohttp
aiosqlite==0.21.0
    # via langgraph-checkpoint-sqlite
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via httpx
attrs==25.4.0
    # via aiohttp
brotli==1.1.0
    # via httpx
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via
    #   ddgs
    #   duckduckgo-search
dataclasses-json==0.6.7
    # via langchain-community
ddgs==9.7.0
    # via -r requirements.in
duckduckgo-search==8.1.1
    # via -r requirements.in
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.16.0
    # via httpcore
h2==4.3.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   ddgs
    #   langgraph-sdk
    #   langsmith
    #   ollama
httpx-sse==0.4.3
    # via langchain-community
hyperframe==6.1.0
    # via h2
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
iniconfig==2.1.0
    # via pytest
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain-classic==1.0.0
    # via langchain-community
langchain-community==0.4.1
    # via -r requirements.in
langchain-core==1.0.2
    # via
    #   -r requirements.in
    #   langchain-classic
    #   langchain-community
    #   langchain-ollama
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-ollama==1.0.0
    # via -r requirements.in
langchain-text-splitters==1.0.0
    # via langchain-classic
langgraph==1.0.2
    # via -r requirements.in
langgraph-checkpoint==3.0.0
    # via
    #   langgraph
    #   langgraph-checkpoint-sqlite
    #   langgraph-prebuilt
langgraph-checkpoint-sqlite==3.0.0
    # via -r requirements.in
langgraph-prebuilt==1.0.2
    # via langgraph
langgraph-sdk==0.2.9
    # via langgraph
langsmith==0.4.39
    # via
    #   -r requirements.in
    #   langchain-classic
    #   langchain-community
    #   langchain-core
lark==1.3.1
    # via -r requirements.in
lxml==6.0.2
    # via
    #   ddgs
    #   duckduckgo-search
marshmallow==3.26.1
    # via dataclasses-json
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
numpy==2.3.4
    # via langchain-community
ollama==0.6.0
    # via
    #   -r requirements.in
    #   langchain-ollama
orjson==3.11.4
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.11.0
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   langchain-core
    #   langsmith
    #   marshmallow
    #   pytest
pluggy==1.6.0
    # via pytest
primp==0.15.0
    # via
    #   ddgs
    #   duckduckgo-search
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pydantic==2.12.3
    # via
    #   langchain-classic
    #   langchain-core
    #   langgraph
    #   langsmith
    #   ollama
    #   pydantic-settings
pydantic-core==2.41.4
    # via pydantic
pydantic-settings==2.11.0
    # via langchain-community
pygments==2.19.2
    # via pytest
pytest==8.4.2
    # via -r requirements.in
python-dotenv==1.2.1
    # via
    #   -r requirements.in
    #   pydantic-settings
pyyaml==6.0.3
    # via
    #   langchain-classic
    #   langchain-community
    #   langchain-core
requests==2.32.5
    # via
    #   langchain-classic
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
sniffio==1.3.1
    # via anyio
socksio==1.0.0
    # via httpx
sqlalchemy==2.0.44
    # via
    #   -r requirements.in
    #   langchain-classic
    #   langchain-community
sqlite-vec==0.1.6
    # via langgraph-checkpoint-sqlite
tenacity==9.1.2
    # via
    #   langchain-community
    #   langchain-core
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   aiosqlite
    #   anyio
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via requests
xxhash==3.6.0
    # via langgraph
yarl==1.22.0
    # via aiohttp
zstandard==0.25.0
    # via langsmith
